Move metacluster into its own directory and static library
