+++ fdbserver/ClusterController.actor.cpp
@@ -50,8 +50,6 @@
 #include "fdbserver/ClusterRecovery.actor.h"
 #include "fdbserver/DataDistributorInterface.h"
 #include "fdbserver/DBCoreState.h"
-#include "fdbclient/Metacluster.h"
-#include "fdbclient/MetaclusterManagement.actor.h"
 #include "fdbserver/MoveKeys.actor.h"
 #include "fdbserver/LeaderElection.h"
 #include "fdbserver/LogSystem.h"
@@ -74,6 +72,9 @@
 #include "flow/Error.h"
 #include "flow/Trace.h"
 #include "flow/Util.h"
+
+#include "metacluster/MetaclusterMetrics.h"
+
 #include "flow/actorcompiler.h" // This must be the last #include.
 
 ACTOR Future<Optional<Value>> getPreviousCoordinators(ClusterControllerData* self) {
@@ -2802,42 +2803,12 @@ ACTOR Future<Void> metaclusterMetricsUpdater(ClusterControllerData* self) {
 		choose {
 			when(wait(self->db.serverInfo->onChange())) {}
 			when(wait(updaterDelay)) {
-				state Reference<ReadYourWritesTransaction> tr = makeReference<ReadYourWritesTransaction>(self->cx);
-				loop {
-					try {
-						tr->setOption(FDBTransactionOptions::READ_SYSTEM_KEYS);
-						state std::map<ClusterName, DataClusterMetadata> clusters;
-						state int64_t tenantCount;
-						wait(store(clusters,
-						           MetaclusterAPI::listClustersTransaction(
-						               tr, ""_sr, "\xff"_sr, CLIENT_KNOBS->MAX_DATA_CLUSTERS)) &&
-						     store(tenantCount,
-						           MetaclusterAPI::ManagementClusterMetadata::tenantMetadata().tenantCount.getD(
-						               tr, Snapshot::False, 0)));
-						state std::pair<ClusterUsage, ClusterUsage> capacityNumbers =
-						    MetaclusterAPI::metaclusterCapacity(clusters);
-
-						MetaclusterMetrics metrics;
-						metrics.numTenants = tenantCount;
-						metrics.numDataClusters = clusters.size();
-						metrics.tenantGroupCapacity = capacityNumbers.first.numTenantGroups;
-						metrics.tenantGroupsAllocated = capacityNumbers.second.numTenantGroups;
-						self->db.metaclusterMetrics = metrics;
-						TraceEvent("MetaclusterCapacity")
-						    .detail("TotalTenants", self->db.metaclusterMetrics.numTenants)
-						    .detail("DataClusters", self->db.metaclusterMetrics.numDataClusters)
-						    .detail("TenantGroupCapacity", self->db.metaclusterMetrics.tenantGroupCapacity)
-						    .detail("TenantGroupsAllocated", self->db.metaclusterMetrics.tenantGroupsAllocated);
-						break;
-					} catch (Error& e) {
-						TraceEvent("MetaclusterUpdaterError").error(e);
-						// Cluster can change types during/before a metacluster transaction
-						// and throw an error due to timing issues.
-						// In such cases, go back to choose loop instead of retrying
-						if (e.code() == error_code_invalid_metacluster_operation) {
-							break;
-						}
-						wait(tr->onError(e));
+				try {
+					wait(store(self->db.metaclusterMetrics, MetaclusterMetrics::getMetaclusterMetrics(self->cx)));
+				} catch (Error& e) {
+					// Ignore errors about the cluster changing type
+					if (e.code() != error_code_invalid_metacluster_operation) {
+						throw;
 					}
 				}
 			}
