+++ metacluster/include/metacluster/MetaclusterMetrics.h
@@ -1,0 +1,39 @@
+/*
+ * MetaclusterMetrics.h
+ *
+ * This source file is part of the FoundationDB open source project
+ *
+ * Copyright 2013-2023 Apple Inc. and the FoundationDB project authors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+#ifndef METACLUSTER_METACLUSTERMETRICS_H
+#define METACLUSTER_METACLUSTERMETRICS_H
+#pragma once
+
+#include "fdbclient/NativeAPI.actor.h"
+#include "flow/flow.h"
+
+struct MetaclusterMetrics {
+	int numTenants = 0;
+	int numDataClusters = 0;
+	int tenantGroupCapacity = 0;
+	int tenantGroupsAllocated = 0;
+
+	MetaclusterMetrics() = default;
+
+	static Future<MetaclusterMetrics> getMetaclusterMetrics(Database db);
+};
+
+#endif
